refactor(table): extract transaction protobuf conversions into its own module
The thirteen conversions between the transaction types and pb::Transaction were
about 700 lines interleaved with the logic they serialize. They are the format
contract for this module -- a field added to an Operation is only durable once
it round-trips here -- so grouping them makes that contract reviewable in one
place.
